requirements.txt
Django==4.0.3
Django==3.2
djangorestframework==3.14.0
joblib==1.2.0
keras==2.12.0
librosa==0.10.0.post2
numpy==1.23.5
pandas==2.0.1
python_dateutil==2.8.2
Requests==2.31.0
